It works

+++ Project/NextItemPredDataset.py
@@ -23,7 +23,8 @@ class NextItemPredDataset(Dataset):
         return (
             self.data[index].user_id,
             self.data[index].items_ids,
-            self.data[index].times,
+            # Convert times to float
+            self.data[index].times.float(),
             self.data[index].pred_index,
             self.data[index].true_item_id,
         )
